fix(agent): catalog parser honors defined:false + unit-scoped test creds
① catalog_errors.py: _ErrorEnvelope gains a `defined` flag (defaults False —
the wire is untrusted); parse_catalog_error builds a typed CatalogError only when
the server marks the error defined:true. A drifted known code (defined:false)
falls back to the status heuristic, so a code the server itself could not vouch
for can no longer be parsed into a trusted typed exception. Test added.

③ Narrow the fake DEEPSEEK/MIMO/SUPABASE credentials + mock settings out of the
shared tests/conftest.py into a unit-only tests/unit/conftest.py, so eval runs
start from a clean environment. Delete the eval_common._ENV_FAKE_SENTINELS
magic-string coupling; real_env_updates() now fills only unset keys (a real
process value always wins). Unit and eval run as separate invocations, so the
narrowing does not pollute eval env resolution. Unit tests updated to the
simplified contract.

Refs #263
