+++ apps/agent/agent/clients/catalog_errors.py
@@ -155,6 +155,7 @@ class UpstreamUnavailableError(CatalogError, TransientAPIError):
 class _ErrorEnvelope(BaseModel):
     """The oRPC error body ``{defined, code, status, message, data}`` (untrusted)."""
 
+    defined: bool = False
     code: str = ""
     message: str = ""
     data: object = None
@@ -210,7 +211,7 @@ def parse_catalog_error(status_code: int, body: object, url: str) -> APIError:
     """
     envelope = _parse_envelope(body)
     builder = _BUILDERS.get(envelope.code) if envelope is not None else None
-    if envelope is None or builder is None:
+    if envelope is None or envelope.defined is not True or builder is None:
         return _status_fallback(status_code, url)
     _log_wire_message(envelope, url)
     return builder(envelope.data)
